requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.1
joblib==1.2.0
scikit-learn==1.3.1
umap-learn==0.5.4
pandas==2.1.0
numpy==1.24.3
pyyaml==6.0.1
requests==2.31.0


